Added Android UI API mod.
I have created a class called GameActivity which clients can implement.
It does a lot of grunt work behind the scenes, and even allows the
client to use the UI as normal. When I'm finished with it, it will
actually draw the UI (other than the GLSurfaceView) to a Bitmap and
then render that Bitmap onto a full-screen quad over the other rendered
primitives. Hopefully this will allow for fast UIs using the standard
framework.